Switch contents between TokenMintForm and TokenMintDialog
