+++ src/modules/dashboard/components/TokenMintDialog/TokenMintDialog.tsx
@@ -1,40 +1,33 @@
 import { FormikProps } from 'formik';
-import React from 'react';
+import React, { useCallback } from 'react';
 import { defineMessages } from 'react-intl';
-import { ColonyRole } from '@colony/colony-js';
+import * as yup from 'yup';
+import { useHistory } from 'react-router-dom';
+import { bigNumberify } from 'ethers/utils';
+import moveDecimal from 'move-decimal-point';
 
-import Button from '~core/Button';
 import Dialog, { DialogProps } from '~core/Dialog';
-import DialogSection from '~core/Dialog/DialogSection';
-import { Annotations, Input } from '~core/Fields';
-import PermissionRequiredInfo from '~core/PermissionRequiredInfo';
-import Heading from '~core/Heading';
-import { Colony, useLoggedInUser } from '~data/index';
+import { ActionForm } from '~core/Fields';
+import { ActionTypes } from '~redux/index';
+import { pipe, mapPayload, withMeta } from '~utils/actions';
+import { Colony } from '~data/index';
 import { getTokenDecimalsWithFallback } from '~utils/tokens';
-import { WizardDialogType, useTransformer } from '~utils/hooks';
-import { getAllUserRoles } from '../../../transformers';
-import { hasRoot } from '../../../users/checks';
+import { WizardDialogType } from '~utils/hooks';
 
 import TokenMintForm from './TokenMintForm';
 
-import styles from './TokenMintDialog.css';
-
 const MSG = defineMessages({
-  title: {
-    id: 'admin.Tokens.TokenMintDialog.dialogTitle',
-    defaultMessage: 'Mint new tokens',
+  errorAmountMin: {
+    id: 'admin.Tokens.TokenMintDialog.errorAmountMin',
+    defaultMessage: 'Please enter an amount greater than 0.',
   },
-  amountLabel: {
-    id: 'admin.Tokens.TokenMintDialog.amountLabel',
-    defaultMessage: 'Amount',
-  },
-  justificationLabel: {
-    id: 'admin.Tokens.TokenMintDialog.amountLabel',
-    defaultMessage: `Explain why you're minting more tokens (optional)`,
+  errorAmountRequired: {
+    id: 'admin.Tokens.TokenMintDialog.errorAmountRequired',
+    defaultMessage: 'Please enter an amount.',
   },
 });
 
-interface FormValues {
+export interface FormValues {
   justification: string;
   mintAmount: number;
 }
@@ -50,99 +43,73 @@ type Props = DialogProps &
 
 const displayName = 'dashboard.TokenMintDialog';
 
+const validationSchema = yup.object().shape({
+  justification: yup.string(),
+  mintAmount: yup
+    .number()
+    .required(() => MSG.errorAmountRequired)
+    .moreThan(0, () => MSG.errorAmountMin),
+});
+
 const TokenMintDialog = ({
-  colony: { nativeTokenAddress, tokens = [], canMintNativeToken },
+  colony: { nativeTokenAddress, tokens = [], colonyAddress, colonyName },
   colony,
   cancel,
   close,
   callStep,
   prevStep,
 }: Props) => {
-  const { walletAddress } = useLoggedInUser();
-
-  const allUserRoles = useTransformer(getAllUserRoles, [colony, walletAddress]);
-
-  const userHasPermissions = canMintNativeToken && hasRoot(allUserRoles);
-  const requiredRoles: ColonyRole[] = [ColonyRole.Root];
+  const history = useHistory();
 
   const nativeToken =
     tokens && tokens.find(({ address }) => address === nativeTokenAddress);
 
+  const transform = useCallback(
+    pipe(
+      mapPayload(({ mintAmount: inputAmount }) => {
+        // Find the selected token's decimals
+        const amount = bigNumberify(
+          moveDecimal(
+            inputAmount,
+            getTokenDecimalsWithFallback(nativeToken?.decimals),
+          ),
+        );
+        return {
+          colonyAddress,
+          colonyName,
+          nativeTokenAddress: nativeToken?.address,
+          amount,
+        };
+      }),
+      withMeta({ history }),
+    ),
+    [],
+  );
+
   return (
-    <Dialog cancel={cancel}>
-      <TokenMintForm
-        colony={colony}
-        nativeToken={nativeToken}
-        onSuccess={close}
-      >
-        {({ handleSubmit, isSubmitting, isValid }: FormikProps<FormValues>) => (
-          <>
-            <DialogSection appearance={{ theme: 'heading' }}>
-              <Heading
-                appearance={{ size: 'medium', margin: 'none', theme: 'dark' }}
-                text={MSG.title}
-              />
-            </DialogSection>
-            {!userHasPermissions && (
-              <DialogSection appearance={{ theme: 'sidePadding' }}>
-                <PermissionRequiredInfo requiredRoles={requiredRoles} />
-              </DialogSection>
-            )}
-            <DialogSection appearance={{ theme: 'sidePadding' }}>
-              <div className={styles.inputContainer}>
-                <div className={styles.inputComponent}>
-                  <Input
-                    appearance={{ theme: 'minimal' }}
-                    formattingOptions={{
-                      numeral: true,
-                      numeralPositiveOnly: true,
-                      numeralDecimalScale: getTokenDecimalsWithFallback(
-                        nativeToken?.decimals,
-                      ),
-                    }}
-                    label={MSG.amountLabel}
-                    name="mintAmount"
-                    disabled={!userHasPermissions}
-                  />
-                </div>
-                <span
-                  className={styles.nativeToken}
-                  title={nativeToken?.name || undefined}
-                >
-                  {nativeToken?.symbol}
-                </span>
-              </div>
-            </DialogSection>
-            <DialogSection appearance={{ theme: 'sidePadding' }}>
-              <div className={styles.annotation}>
-                <Annotations
-                  label={MSG.justificationLabel}
-                  name="annotation"
-                  disabled={!userHasPermissions}
-                />
-              </div>
-            </DialogSection>
-            <DialogSection appearance={{ align: 'right', theme: 'footer' }}>
-              <Button
-                appearance={{ theme: 'secondary', size: 'large' }}
-                onClick={
-                  prevStep && callStep ? () => callStep(prevStep) : undefined
-                }
-                text={{ id: 'button.back' }}
-              />
-              <Button
-                appearance={{ theme: 'primary', size: 'large' }}
-                onClick={() => handleSubmit()}
-                text={{ id: 'button.confirm' }}
-                loading={isSubmitting}
-                disabled={!isValid || !userHasPermissions}
-                style={{ width: styles.wideButton }}
-              />
-            </DialogSection>
-          </>
-        )}
-      </TokenMintForm>
-    </Dialog>
+    <ActionForm
+      initialValues={{
+        justification: '',
+        mintAmount: 0,
+      }}
+      validationSchema={validationSchema}
+      submit={ActionTypes.COLONY_ACTION_MINT_TOKENS}
+      error={ActionTypes.COLONY_ACTION_MINT_TOKENS_ERROR}
+      success={ActionTypes.COLONY_ACTION_MINT_TOKENS_SUCCESS}
+      onSuccess={close}
+      transform={transform}
+    >
+      {(formValues: FormikProps<FormValues>) => (
+        <Dialog cancel={cancel}>
+          <TokenMintForm
+            {...formValues}
+            colony={colony}
+            back={prevStep && callStep ? () => callStep(prevStep) : undefined}
+            nativeToken={nativeToken}
+          />
+        </Dialog>
+      )}
+    </ActionForm>
   );
 };
